backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
django-filter==25.2
psycopg2-binary==2.9.12
python-dotenv==1.2.2
Pillow==12.2.0
